% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_econ_series.R
\name{get_econ_series}
\alias{get_econ_series}
\title{Accessing FRED
This function can pull time series from the FRED database: \url{https://fred.stlouisfed.org}}
\usage{
get_econ_series(series_id,
series_name = NULL,observation_start = NULL,
 observation_end = NULL)
}
\arguments{
\item{series_id}{FRED times series ID, the unique id retrieved from fed database.}

\item{series_name}{Choose a name for the series column in output. Default: series_id.}

\item{observation_start}{Choose a name for the series column in output. Default: series_id.}

\item{observation_end}{Date of last observation in "yyyy-mm-dd" format. Default: Last observation available.}
}
\description{
Accessing FRED
This function can pull time series from the FRED database: \url{https://fred.stlouisfed.org}
}
\examples{
get_econ_series("UNRATE", "unemploymentRate")
}
