% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fedeconseries.R
\name{get_fedeconseries}
\alias{get_fedeconseries}
\title{US federal reserve economic data series accessing
This wrapper function is to access data from  \url{https://alfred.stlouisfed.org}.}
\usage{
get_fedeconseries(s_id, s_name = NULL,
obs_start = NULL, obs_end = NULL,real_start = NULL, real_end = NULL)
}
\arguments{
\item{s_id}{econimic variable names provided by ALFRED, unique identifier.}

\item{s_name}{Output column name, default is s_id.}

\item{obs_start}{Date of first observation in "yyyy-mm-dd" format. Default: Earliest observation available.}

\item{obs_end}{Date of last observation in "yyyy-mm-dd" format. Default: Last observation available.}

\item{real_start}{Date of first real time period in "yyyy-mm-dd" format. Default: First vintage date available.}

\item{real_end}{Date of last real time period in "yyyy-mm-dd" format. Default: Last vintage date available.}
}
\description{
US federal reserve economic data series accessing
This wrapper function is to access data from  \url{https://alfred.stlouisfed.org}.
}
\details{
fromJSON return: (1) $observations(4 cols: realtime_start, realtime_end, date, value); (2)other attrs: $realtime_start, $realtime_end, $observation_start, $observation_end, $units, $output_type, $file_type, $order_by, $sort_order, $count, $offset, $limit
}
\examples{
\dontrun{
    get_fedeconseries("INDPRO", "indpro")
    }

get_fedeconseries("INDPRO", "indpro", real_start = "2010-01-01", real_end = "2015-12-31")
}
\keyword{US}
\keyword{alfred;}
\keyword{federal}
\keyword{reserve}
