% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Federal Reserve Econnomic Series Data}
\usage{
hello(x)
}
\arguments{
\item{x}{the input, the name you want to say hello to}
}
\value{
the output from \code{\link{print}}
}
\description{
To obtain US federal reserve economc data.  Reserach use only.
}
\examples{
hello("John")
\dontrun{
hello("Steve")
}





}
