% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printer.R
\name{printer}
\alias{printer}
\title{printer dunction}
\usage{
printer(r, x)
}
\arguments{
\item{r}{the second column}

\item{x}{the first column}
}
\value{
a tibble
}
\description{
printer dunction
}
\examples{
printer(x=rnorm(5), r=rnorm(5))
}
