% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_fedeconseries  wrapper function name}
\alias{search_fedeconseries  wrapper function name}
\alias{search_fedeconseries}
\title{Get US federal reserve economic data series that match keywords
This wrapper function is to access data from  \url{https://alfred.stlouisfed.org}}
\usage{
search_fedeconseries(search_text, search_type = 'full_text',
                            realtime_start = NULL, realtime_end = NULL)
}
\arguments{
\item{search_text}{The words to match against economic data series.}

\item{search_type}{Determines the type of search to perform.
One of the following strings: 'full_text', 'series_id'.}

\item{realtime_start}{Date of first real time period in "yyyy-mm-dd" format..}

\item{realtime_end}{Date of last real time period in "yyyy-mm-dd" format.}
}
\value{
json file
}
\description{
Get US federal reserve economic data series that match keywords
This wrapper function is to access data from  \url{https://alfred.stlouisfed.org}
}
\examples{
search_fedeconseries('GDP', search_type = 'full_text',realtime_start = NULL, realtime_end = NULL)

}
